requirements.txt
# requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch3d>=0.7.0
diffusers>=0.21.0
transformers>=4.35.0
streamlit>=1.28.0
plotly>=5.14.0
numpy>=1.24.0
open3d>=0.17.0
trimesh>=3.23.0
pillow>=10.0.0
scipy>=1.10.0
matplotlib>=3.7.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
opencv-python>=4.8.0
pytorch-lightning>=2.0.0
einops>=0.7.0
accelerate>=0.24.0
xformers>=0.0.22
omegaconf>=2.3.0
torch-ema>=0.3.0
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
docker>=6.1.0